requirements.txt
playwright>=1.49.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pydantic>=2.5.0 # Optional, but good practice
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
